forecasting/requirements-macos.txt
# Database
mysqlclient==2.2.0
SQLAlchemy==2.0.23

# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# Deep Learning (macOS specific)
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
keras==2.15.0
keras-tcn==3.5.0

# Data Processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0

